server/requirements.txt
fastapi
uvicorn
boto3
pydantic-settings
python-dotenv
sqlmodel
psycopg2-binary
redis
minio
python-multipart
alembic